% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_data.R
\name{create_random_data}
\alias{create_random_data}
\title{Create random data set with 4 clusters in a 2 dimensional subspace of a nr_othter_vars+2 dimensional space}
\usage{
create_random_data(n = 10000, nr_other_vars = 4)
}
\arguments{
\item{n}{number of points}

\item{nr_other_vars}{number of other variables / "noise" dimensions}
}
\value{
list containing the random data.table and a vector with the true underlying cluster assignments
}
\description{
Create random data set with 4 clusters in a 2 dimensional subspace of a nr_othter_vars+2 dimensional space
}
\examples{
reate_random_data(n=1e3)

}
