% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureImpCluster-package.R
\docType{package}
\name{FeatureImpCluster-package}
\alias{FeatureImpCluster}
\alias{FeatureImpCluster-package}
\title{FeatureImpCluster: Feature importance for partitional clustering}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Oliver Pfaffel \email{opfaffel@posteo.net}

}
\keyword{internal}
