% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureimpcluster.R
\name{plot.featImpCluster}
\alias{plot.featImpCluster}
\title{Feature importance box plot}
\usage{
\method{plot}{featImpCluster}(x, dat = NULL, color = "none", showPoints = FALSE, ...)
}
\arguments{
\item{x}{an object returned from FeatureImpCluster}

\item{dat}{same data as used for the computation of the feature importance (only relevant for colored plots)}

\item{color}{If set to "type", the plot will show different variable types with a different color.}

\item{showPoints}{Show points (default is False)}

\item{...}{arguments to be passed to base plot method}
}
\value{
Returns a ggplot2 object
}
\description{
Feature importance box plot
}
